Store modrm offset instead of the byte it self.
This is once again remnants of the state machine, where the modrm byte
had to be passed into sproc_grpfixup. An offset is more useful for the
user, since from it they can also access the sib byte, and the memory
offset.

Added an _off suffix so that it's obvious that these are offsets and
not the actual values.