Error on VEX/EVEX/XOP
We pretend we're from 2008 and treat these prefixes as #UD.
Detecting every possible #UD sequence was never a goal for this lib,
so in many cases we just allow instruction aliasing to happen, for
simplicity. Here, however, we do know that aliasing is not the right way
to parse these instructions, so instead of pretending like we do know
how to do it, let's just throw an error.